+++ packages/ui/src/components/Table/definition.ts
@@ -27,6 +27,20 @@ import type {
 } from './types';
 import styles from './Table.module.css';
 
+/** @internal */
+export const TableWrapperDefinition = defineComponent<{
+  className?: string;
+}>()({
+  styles,
+  classNames: {
+    root: 'bui-TableWrapper',
+    resizableContainer: 'bui-TableResizableContainer',
+  },
+  propDefs: {
+    className: {},
+  },
+});
+
 /**
  * Component definition for Table
  * @public
@@ -38,6 +52,8 @@ export const TableDefinition = defineComponent<TableRootOwnProps>()({
   },
   propDefs: {
     stale: { dataAttribute: true },
+    isPending: { dataAttribute: true },
+    loading: { dataAttribute: true },
   },
 });
 
@@ -76,6 +92,7 @@ export const TableBodyDefinition = defineComponent<TableBodyOwnProps>()({
 export const RowDefinition = defineComponent<RowOwnProps>()({
   styles,
   analytics: true,
+  bg: 'consumer',
   classNames: {
     root: 'bui-TableRow',
     cell: 'bui-TableCell',